% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{infer_health}
\alias{infer_health}
\title{Check API health}
\usage{
infer_health(client)
}
\arguments{
\item{client}{An infer_client object}
}
\value{
A list with API status information
}
\description{
Check API health
}
