% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{infer_model_config}
\alias{infer_model_config}
\title{Get optimal inference configuration for a model}
\usage{
infer_model_config(client, model_id, n_texts = 100)
}
\arguments{
\item{client}{An infer_client object}

\item{model_id}{The model identifier}

\item{n_texts}{Expected number of texts (affects parallel recommendation)}
}
\value{
A list with batch_size, device_mode, use_parallel, training_mode, etc.
}
\description{
Get optimal inference configuration for a model
}
