% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{infer_model_info}
\alias{infer_model_info}
\title{Get model information}
\usage{
infer_model_info(client, model_id)
}
\arguments{
\item{client}{An infer_client object}

\item{model_id}{The model identifier}
}
\value{
A list with model metadata
}
\description{
Get model information
}
