% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{infer_models}
\alias{infer_models}
\title{List available models}
\usage{
infer_models(client)
}
\arguments{
\item{client}{An infer_client object}
}
\value{
A list of available models with training_mode and multi_label info
}
\description{
List available models
}
