% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{infer_resources}
\alias{infer_resources}
\title{Get server resources}
\usage{
infer_resources(client)
}
\arguments{
\item{client}{An infer_client object}
}
\value{
A list with CPU, GPU, memory, storage, and capacity info
}
\description{
Get server resources
}
