% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{infer_translate_languages}
\alias{infer_translate_languages}
\title{List TranslateGemma supported languages}
\usage{
infer_translate_languages(client)
}
\arguments{
\item{client}{An infer_client object}
}
\value{
A list with 'languages' (list of code/name pairs) and 'count'
}
\description{
Get the list of all languages supported by TranslateGemma.
}
\examples{
\dontrun{
client <- infer_connect("https://your-server.example.com", "YOUR_API_KEY")
langs <- infer_translate_languages(client)
print(paste(langs$count, "languages supported"))
}
}
